ml_anomaly_detection/requirements.txt
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
kafka-python==2.0.2